requirements.txt
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
Flask==3.0.3
idna==3.7
imagekitio==4.0.1
importlib_metadata==8.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
python-dotenv==1.0.1
requests==2.32.3
requests-toolbelt==0.10.1
urllib3==1.26.19
Werkzeug==3.0.3
zipp==3.19.2
